pyproject.toml
[tool.poetry]
name = "lsu-pilot"
version = "0.1.0"
description = ""
authors = ["Shawn Michael Larkin <shawn.michael.larkin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.7"
python-telegram-bot = "^21.1.1"
openai = "^1.25.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
